HtmlIndex API separated, release mods updated not to use auto-deps.
