+++ ide/html.indexing/manifest.mf
@@ -1,0 +1,6 @@
+Manifest-Version: 1.0
+AutoUpdate-Show-In-Client: false
+OpenIDE-Module: org.netbeans.modules.html.indexing
+OpenIDE-Module-Localizing-Bundle: org/netbeans/modules/html/indexing/Bundle.properties
+OpenIDE-Module-Specification-Version: 1.0
+
