+++ nbbuild/cluster.properties
@@ -402,6 +402,7 @@ nb.cluster.ide=\
         html.custom,\
         html.editor,\
         html.editor.lib,\
+        html.indexing,\
         html.lexer,\
         html.parser,\
         html.validation,\
