+++ webcommon/javascript2.nodejs/nbproject/project.xml
@@ -79,15 +79,6 @@
                     </run-dependency>
                 </dependency>
                 <dependency>
-                    <code-name-base>org.netbeans.modules.html.editor</code-name-base>
-                    <build-prerequisite/>
-                    <compile-dependency/>
-                    <run-dependency>
-                        <release-version>2</release-version>
-                        <specification-version>2.50</specification-version>
-                    </run-dependency>
-                </dependency>
-                <dependency>
                     <code-name-base>org.netbeans.modules.javascript2.editor</code-name-base>
                     <build-prerequisite/>
                     <compile-dependency/>
